requirements.txt
Flask>=2.3.0
pandas>=1.5.0
pyunpack>=0.3.0
langchain>=0.0.267
langchain-openai>=0.0.1
openai>=0.28.0
langchain-experimental>=0.0.10
tabulate>=0.9.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
numpy>=1.25.0
patool>=1.12
openpyxl>=3.1.0